did that fix it? (still should use the EA predicates)

+++ src/geometric.h
@@ -17,7 +17,7 @@ class Point2
 {
     public:
         Point2() : x(0), y(0) { }
-        Point2(double px, double py) : x(px), y(py) { }
+        Point2(long double px, double py) : x(px), y(py) { }
         ~Point2() { }
         Point2& operator=(const Point2& p) {
             x = p.x;
@@ -63,8 +63,8 @@ class Point2
         }
 
 
-        double x; // cartesian x coordinate
-        double y; // cartesian y coordinate
+        long double x; // cartesian x coordinate
+        long double y; // cartesian y coordinate
 };
 
 inline std::ostream& operator<<(std::ostream& strm, const Point2& a) {
@@ -110,17 +110,17 @@ class Triangle2
 /*
  * distance between 2 points in 2d
  */
-inline double distance(const Point2& p1, const Point2& p2) 
+inline long double distance(const Point2& p1, const Point2& p2) 
 {
-    double a = (p1.x - p2.x);
-    double b = (p1.y - p2.y);
+    long double a = (p1.x - p2.x);
+    long double b = (p1.y - p2.y);
     return sqrt(a*a + b*b);
 }
 
 /*
  * abc are: ccw >0, colinear =0, cw <0
  */
-inline double orient2d(const Point2& a, const Point2& b, const Point2& c)
+inline long double orient2d(const Point2& a, const Point2& b, const Point2& c)
 {
     return (a.x - c.x) * (b.y - c.y) - (b.x - c.x) * (a.y - c.y);
 }
@@ -128,18 +128,18 @@ inline double orient2d(const Point2& a, const Point2& b, const Point2& c)
 /*
  * d is, with circle abc: inside >0, cocircular =0, outside <0  
  */
-inline double incircle(const Point2& a, const Point2& b, const Point2& c, 
+inline long double incircle(const Point2& a, const Point2& b, const Point2& c, 
                        const Point2& d)
 {
-    double amx = a.x - d.x;
-    double amy = a.y - d.y;
-    double bmx = b.x - d.x;
-    double bmy = b.y - d.y;
-    double cmx = c.x - d.x;
-    double cmy = c.y - d.y;
-    double am2 = amx * amx + amy * amy;
-    double bm2 = bmx * bmx + bmy * bmy;
-    double cm2 = cmx * cmx + cmy * cmy;
+    long double amx = a.x - d.x;
+    long double amy = a.y - d.y;
+    long double bmx = b.x - d.x;
+    long double bmy = b.y - d.y;
+    long double cmx = c.x - d.x;
+    long double cmy = c.y - d.y;
+    long double am2 = amx * amx + amy * amy;
+    long double bm2 = bmx * bmx + bmy * bmy;
+    long double cm2 = cmx * cmx + cmy * cmy;
     return amx*bmy*cm2 + cmx*amy*bm2 + bmx*cmy*am2 - 
            bmx*amy*cm2 - amx*cmy*bm2 - cmx*bmy*am2;
 }
